Imporving install.sh to work on mac

+++ install.sh
@@ -1,47 +1,65 @@
 #!/bin/bash 
-
+OS=''
 function CurrentSystem()
 {
-    # Checking the underlying operating system using the lsb_release (Linux Standard Base) command and others
     echo "#####################################"
-    echo ">>># Checking which Linux system is running" 
-    if [ -f "/etc/debian_version" ]
-    then
-        echo "### This Linux system is a debian-based system" 
-        echo ">>># Checking which debian flavor your system is" 
-        if [ "$(grep -i 'ubuntu' /etc/*release)" ]; 
+    echo $'>>># Detecting the underlying OS'
+    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        if [ -f "/etc/debian_version" ]
         then
-            ubuntuVersion=$(lsb_release -r 2> /dev/null | sed 's/:\t/:/' | cut -d ':' -f 2-)
-            echo "### This Linux system is an ubuntu-based system" $ubuntuVersion
+            OS='DebianBased'
+        else 
+            OS='RedHatBased'
         fi
-    else 
-        echo "***Error: This project expect a debian-based operating systems"
-        SystemFlavor=$(cat /etc/issue | awk '{print $1}')
-        echo "You system is "$SystemFlavor
-        exit 1
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+            OS='MacOS'
     fi
 }
 
-function CurrentPython
+function InstallDependencies
 {
     echo "#####################################"
     echo $'>>># Checking Python requirement'
     PythonVersion="$(python3 -V 2>&1 | awk '{print $2}' | awk -F. '{print $1}')"  
     if [[ PythonVersion -eq 3 ]]
     then
-        
         echo "Python3 requirement already satisfied"
     else
-        echo "Installing python 3"    
-        sudo apt-get update &> /dev/null 
-        sudo apt-get install python3.6 &> /dev/null
-        if [ $? -eq 0 ]
+        echo "Installing python 3" 
+        if [[ "$OS" == "DebianBased" ]]
+        then
+            echo "deb"
+            sudo apt-get update &> /dev/null 
+            sudo apt-get install python3.6 &> /dev/null
+            if [ $? -eq 0 ]
+            then 
+                echo "Python 3.6 has been installed successfully"
+            else
+                echo "An Error occured while installing python 3.6"                
+                exit 1
+            fi
+        elif [[ "$OS" == "RedHatBased" ]]
         then 
-            echo "Python 3.6 has been installed successfully"
-        else 
-            echo "An Error occured while installing python 3.6" 
-            echo "Check your internet connectivity"
-            exit 1
+            sudo yum update &> /dev/null 
+            sudo yum install python3.6 &> /dev/null
+            if [ $? -eq 0 ]
+            then 
+                echo "Python 3.6 has been installed successfully"
+            else 
+                echo "An Error occured while installing python 3.6"                
+                exit 1
+            fi
+        elif [[ "$OS" == "MacOS"  ]]
+        then 
+            brew install python3.6 &> /dev/null
+            if [ $? -eq 0 ]
+            then 
+                echo "Python 3.6 has been installed successfully"
+            else 
+                echo "An Error occured while installing python 3.6"                
+                exit 1
+            fi
+            rehash
         fi
     fi  
 }
@@ -51,27 +69,58 @@ function InstallingBinwalk
     python3 -c "import binwalk" 2> /dev/null
     if [ $? -eq 0 ]
     then 
-        echo "binwalk python 3 module is already installed"
+        echo "binwalk module is already installed"
     else 
-
-        echo "#####################################"
-        echo $'### Installing binwalk python 3 module'
-        sudo apt-get install git > /dev/null
-        sudo rm -rf /tmp/binwalk
-        git clone https://github.com/devttys0/binwalk /tmp/binwalk  2> /dev/null
-        cd /tmp/binwalk
-        sudo python3 /tmp/binwalk/setup.py install > /dev/null
-        python3 -c "import binwalk" 2> /dev/null
-        if [ $? -eq 0 ]
+        echo "Installing binwalk" 
+        if [[ "$OS" == "DebianBased" ]]
         then
-            echo "The binwalk module has been successfully installed" 
-        else
-            echo "***Error: The binwalk  wasn't installed"
-            exit 1
-        fi 
+            sudo apt-get install git > /dev/null
+            sudo rm -rf /tmp/binwalk
+            git clone https://github.com/devttys0/binwalk /tmp/binwalk  2> /dev/null
+            cd /tmp/binwalk
+            sudo python3 /tmp/binwalk/setup.py install > /dev/null
+            python3 -c "import binwalk" 2> /dev/null
+            if [ $? -eq 0 ]
+            then
+                echo "The binwalk module has been successfully installed" 
+            else
+                echo "Error: The binwalk  wasn't installed"
+                exit 1
+            fi 
+        elif [[ "$OS" == "RedHatBased" ]]
+        then 
+            sudo yum install git > /dev/null
+            sudo rm -rf /tmp/binwalk
+            git clone https://github.com/devttys0/binwalk /tmp/binwalk  2> /dev/null
+            cd /tmp/binwalk
+            sudo python3 /tmp/binwalk/setup.py install > /dev/null
+            python3 -c "import binwalk" 2> /dev/null
+            if [ $? -eq 0 ]
+            then
+                echo "The binwalk module has been successfully installed" 
+            else
+                echo "Error: The binwalk  wasn't installed"
+                exit 1
+            fi 
+        elif [[ "$OS" == "MacOS"  ]]
+        then 
+            brew install git > /dev/null
+            sudo rm -rf /tmp/binwalk
+            git clone https://github.com/devttys0/binwalk /tmp/binwalk  2> /dev/null
+            cd /tmp/binwalk
+            sudo python3 /tmp/binwalk/setup.py install > /dev/null
+            python3 -c "import binwalk" 2> /dev/null
+            if [ $? -eq 0 ]
+            then
+                echo "The binwalk module has been successfully installed" 
+            else
+                echo "Error: The binwalk  wasn't installed"
+                exit 1
+            fi 
+        fi
     fi
 }
 
 CurrentSystem
-CurrentPython
+InstallDependencies
 InstallingBinwalk
